Address, Owner API handling

+++ src/main/java/com/example/springwithsql/Database/Repository/AddressRepository.java
@@ -3,4 +3,4 @@ package com.example.springwithsql.Database.Repository;
 import com.example.springwithsql.Database.Entity.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressRepository extends JpaRepository<Address, Integer> { }
+public interface AddressRepository extends JpaRepository<Address, Long> { }
